Refactor to centralize state in stardata
