+++ apps/web/src/app/chart-your-stars-demo/mockData.ts
@@ -1,0 +1,98 @@
+// Mock data for demo
+// Alice: '/uploads/user-photos/cmeg4p6r70002ihd1zt0im435.1757255300867.small.webp'
+// Bob: '/uploads/user-photos/cmeg4pewy0000ihkwmy8voxi9.1759609763448.small.webp'
+// Carol: '/uploads/user-photos/cmeg5liai0003ihzcqi2ppfwd.1755611564561.small.webp'
+// David: '/uploads/user-photos/cmeimf6010005ygjemuhgjxqn.1755612613185.small.webp'
+
+export const MOCK_PEOPLE = [
+  {
+    id: '1',
+    name: 'Alice Johnson',
+    photo:
+      '/uploads/user-photos/cmeg4p6r70002ihd1zt0im435.1757255300867.small.webp',
+  },
+  {
+    id: '2',
+    name: 'Bob Smith',
+    photo:
+      '/uploads/user-photos/cmeg4pewy0000ihkwmy8voxi9.1759609763448.small.webp',
+  },
+  {
+    id: '3',
+    name: 'Carol Williams',
+    photo:
+      '/uploads/user-photos/cmeg5liai0003ihzcqi2ppfwd.1755611564561.small.webp',
+  },
+  {
+    id: '4',
+    name: 'David Brown',
+    photo:
+      '/uploads/user-photos/cmeimf6010005ygjemuhgjxqn.1755612613185.small.webp',
+  },
+  {
+    id: '5',
+    name: 'Eve Davis',
+    photo:
+      '/uploads/user-photos/cmeimfkgh0007ygjez1dzqzra.1755612631924.small.webp',
+  },
+  {
+    id: '6',
+    name: 'Frank Miller',
+    photo:
+      '/uploads/user-photos/cmeq3icdd0000ihlv51grw0u6.1756070999426.small.webp',
+  },
+  {
+    id: '7',
+    name: 'Grace Wilson',
+    photo:
+      '/uploads/user-photos/cmeq3j6u10002ihlvwvdlqvmh.1756071046858.small.webp',
+  },
+  {
+    id: '8',
+    name: 'Henry Moore',
+    photo:
+      '/uploads/user-photos/cmeq3jqr70004ihlvvqwwt7kd.1756071078346.small.webp',
+  },
+  {
+    id: '9',
+    name: 'Ivy Taylor',
+    photo:
+      '/uploads/user-photos/cmeq3kbvr0006ihlvvxqxqxqx.1756071109346.small.webp',
+  },
+  {
+    id: '10',
+    name: 'Jack Anderson',
+    photo:
+      '/uploads/user-photos/cmeq3kvvr0008ihlvvyqyqyqy.1756071140346.small.webp',
+  },
+  {
+    id: '11',
+    name: 'Karen Thomas',
+    photo:
+      '/uploads/user-photos/cmeq3lfvr000aihlvvzqzqzqz.1756071171346.small.webp',
+  },
+  {
+    id: '12',
+    name: 'Leo Jackson',
+    photo:
+      '/uploads/user-photos/cmeq3lvvr000cihlvv0q0q0q0.1756071202346.small.webp',
+  },
+  {
+    id: '13',
+    name: 'Mia White',
+    photo:
+      '/uploads/user-photos/cmeq3mfvr000eihlvv1q1q1q1.1756071233346.small.webp',
+  },
+  {
+    id: '14',
+    name: 'Noah Harris',
+    photo:
+      '/uploads/user-photos/cmeq3mvvr000gihlvv2q2q2q2.1756071264346.small.webp',
+  },
+  {
+    id: '15',
+    name: 'Olivia Martin',
+    photo:
+      '/uploads/user-photos/cmeq3nfvr000iihlvv3q3q3q3.1756071295346.small.webp',
+  },
+]
